VW-633 migrate inputtype base on old structure

+++ Services/Accountancy/Services.Accountancy.Service/Category/Implement/CategoryDataService.cs
@@ -138,7 +138,7 @@ namespace VErp.Services.Accountancy.Service.Category
 
             var id = await _accountancyContext.InsertDataTable(dataTable);
             await _activityLogService.CreateLog(EnumObjectType.Category, id, $"Thêm mới dữ liệu danh mục {id}", data.JsonSerialize());
-            return id;
+            return (int)id;
         }
